requirements.txt
flask
python-pptx
pandas
Pillow
